Udacity-Capstone/requirements.txt
alembic==1.11.0
aniso8601==6.0.0
atomicwrites==1.4.1
attrs==22.2.0
aws-shell==0.2.2
awscli==1.27.120
boto3==1.26.120
botocore==1.29.120
cffi==1.15.1
click==8.1.3
colorama==0.4.4
configobj==5.0.8
distlib==0.3.6
docutils==0.16
ecdsa==0.18.0
filelock==3.9.0
Flask==1.1.2
Flask-Cors==3.0.7
Flask-Migrate==4.0.4
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
gunicorn==20.0.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==2.11.3
jmespath==1.0.
Mako==1.2.4
MarkupSafe==1.1.1
packaging==23.1
platformdirs==3.1.0
pluggy==0.13.1
prompt-toolkit==1.0.18
psycopg2-binary==2.8.2
py==1.11.0
pyasn1==0.5.0
pycparser==2.21
Pygments==2.15.1
PyJWT==1.7.1
pytest==6.2.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
pytz==2019.1
PyYAML==5.4.1
rsa==4.7.2
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.12.0
SQLAlchemy==1.3.4
toml==0.10.2
typing-extensions==4.5.0
urllib3==1.26.15
virtualenv==20.20.0
wcwidth==0.2.6
Werkzeug==2.0.3
zipp==3.15.0
